Rename scheduler disaggregation init mixin
